requirements.txt
crf==0.0.1a1
einops==0.3.2
load==2020.12.3
matplotlib==3.4.2
numpy==1.20.3
Pillow==8.4.0
pydensecrf==1.0rc3
scikit_learn==1.0.1
segmentation_models_pytorch==0.2.0
torch==1.7.0
torchvision==0.8.1
